+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/console.php
@@ -26,6 +26,7 @@ use Symfony\Bundle\FrameworkBundle\Command\ContainerDebugCommand;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerLintCommand;
 use Symfony\Bundle\FrameworkBundle\Command\DebugAutowiringCommand;
 use Symfony\Bundle\FrameworkBundle\Command\EventDispatcherDebugCommand;
+use Symfony\Bundle\FrameworkBundle\Command\RateLimiterDebugCommand;
 use Symfony\Bundle\FrameworkBundle\Command\RouterDebugCommand;
 use Symfony\Bundle\FrameworkBundle\Command\RouterMatchCommand;
 use Symfony\Bundle\FrameworkBundle\Command\SecretsDecryptToLocalCommand;
@@ -152,6 +153,12 @@ return static function (ContainerConfigurator $container) {
         ->set('console.command.event_dispatcher_debug', EventDispatcherDebugCommand::class)
             ->args([
                 tagged_locator('event_dispatcher.dispatcher', 'name'),
+            ])
+            ->tag('console.command')
+
+        ->set('console.command.rate_limiter_debug', RateLimiterDebugCommand::class)
+            ->args([
+                tagged_locator('rate_limiter', 'name'),
             ])
             ->tag('console.command')